32: complete update, delete, get single post feature

+++ crud/src/routes/posts/posts.controller.ts
@@ -4,7 +4,7 @@ import { PostsService } from 'src/routes/posts/posts.service'
 import { AuthType, ConditionGuard } from 'src/shared/constants/auth.constant'
 import { ActiveUser } from 'src/shared/decorators/active-user.decorator'
 import { Auth } from 'src/shared/decorators/auth.decorator'
-import { GetPostItemDTO } from './post.dto'
+import { CreatePostBodyDTO, GetPostItemDTO, UpdatePostBodyDTO } from './post.dto'
 
 // import { AccessTokenGuard } from 'src/shared/guards/access-token.guard'
 // import { APIKeyGuard } from 'src/shared/guards/api-key.guard'
@@ -21,22 +21,24 @@ export class PostsController {
 
   @Post()
   @Auth([AuthType.Bearer])
-  createPost(@Body() body: any, @ActiveUser('userId') userId: number) {
-    return this.postsService.createPost(body, userId)
+  async createPost(@Body() body: CreatePostBodyDTO, @ActiveUser('userId') userId: number) {
+    return new GetPostItemDTO(await this.postsService.createPost(body, userId))
   }
 
   @Get(':id')
-  getPost(@Param('id') id: string) {
-    return this.postsService.getPost(id)
+  async getPost(@Param('id') id: string) {
+    return new GetPostItemDTO(await this.postsService.getPost(Number(id)))
   }
 
   @Put(':id')
-  updatePost(@Param('id') id: string, @Body() body: any) {
-    return this.postsService.updatePost(id, body)
+  @Auth([AuthType.Bearer])
+  async updatePost(@Param('id') id: string, @Body() body: UpdatePostBodyDTO, @ActiveUser('userId') userId: number) {
+    return new GetPostItemDTO(await this.postsService.updatePost({ postId: Number(id), body, userId }))
   }
 
   @Delete(':id')
-  deletePost(@Param('id') id: string) {
-    return this.postsService.deletePost(id)
+  @Auth([AuthType.Bearer])
+  async deletePost(@Param('id') id: string, @ActiveUser('userId') userId: number): Promise<boolean> {
+    return await this.postsService.deletePost({ postId: Number(id), userId })
   }
 }
